Remove AWS api bridge
This has been a discussion point for a while. The (mostly generated)
code for the AWS api bridge is by far the largest source component in
Cloudstack, while seldomly used.

Now that alternate options exist to provide EC2 compatibility, it
makes sense to remove it for the few users who cannot directly
talk to the cloudstack API.

Conflicts:
	awsapi/pom.xml
	awsapi/src/com/cloud/bridge/io/S3FileSystemBucketAdapter.java
	awsapi/src/com/cloud/bridge/model/MHostVO.java
	awsapi/src/com/cloud/bridge/persist/dao/MultiPartUploadsDaoImpl.java
	awsapi/src/com/cloud/bridge/service/EC2SoapServiceImpl.java
	awsapi/src/com/cloud/bridge/service/controller/s3/S3BucketAction.java
	awsapi/src/com/cloud/bridge/service/core/ec2/EC2Engine.java
	awsapi/src/com/cloud/bridge/util/EC2RestAuth.java
	awsapi/src/com/cloud/stack/models/ApiConstants.java
	framework/db/src/com/cloud/utils/db/TransactionLegacy.java
	framework/db/test/db.properties
	packaging/centos63/cloud.spec
	plugins/network-elements/globodns/test/resources/db.properties
	pom.xml
	tools/eclipse/eclipse.epf