Allow sysvm configuration to be passed thru the socket when using recent qemu version
It seems that behaviour of virtio "serial" ports has changed a bit with recent qemu version (2.1, 2.2). We now need to shutdown the socket before closing it in order to allow some time to qemu to pass it to the guest and acknowledge back. Without it the socket is being closed before qemu could communicate to the guest and therefore the configuration never get delivered.