live migration: cancel migration job on timeout exception
Currently if a migration job takes longer than the max time for a CS job,
then the disk at the remote is deleted but the job keeps running. This
leads to a situation were the VM could be moved on the destination
hypervisors but without a visible disk on the filesystem.

This fix attempt to cancel the migration job on the source host before
requesting the disk removal at the destination.