Handle associate and addIpToNic the exoscale way
Introducing a new state and field in the user_ip_address table: state: Associated
field: associated

If a new ip is associated, the state of the IP is set to associated and the associated field is set to the date of the association.

We need the associated field to be able to double check the usage of the IP addresses, otherwise we would have it only in the events.

The decision for the new state is based on having a clear status for each of the IP addresses, instead the use of the combination of multiple fields.

DeployVM: Fetch an IP with state Free, set to state Allocated, set allocated date
AssociateIP: Fetch an IP with state Free, set to state Associated, set associated date
AddIpToNic: Fetch an IP with state Associated, set to state Allocated, set allocated date
RemoveIpFromNic: Set IP to state Associated, remove allocated date
DisassociatedIP: Set IP to state Free, remove associated date